refactor: Modularize JavaScript codebase
Refactored the monolithic js/main.js into multiple modules to improve code organization and maintainability.

- Created separate modules for configuration, state, UI, powers, cataclysms, drawing, simulation, and universe logic.
- Updated main.js to be the main entry point, importing from the new modules.
- Updated README.md with the new file structure and instructions for running the project with a local server.
