+++ jules-scratch/verification/verify_refactor.py
@@ -1,0 +1,15 @@
+from playwright.sync_api import sync_playwright
+
+def run():
+    with sync_playwright() as p:
+        browser = p.chromium.launch()
+        page = browser.new_page()
+
+        page.goto('http://localhost:8000')
+        # Wait for the particles to be rendered
+        page.wait_for_selector('#particles-js canvas')
+        page.screenshot(path='jules-scratch/verification/verification.png')
+        browser.close()
+
+if __name__ == '__main__':
+    run()
